feat(service) implement INewsService interface and refactor NewsService for better validation

+++ NewsAI.Data/Service/NewsService.cs
@@ -4,11 +4,9 @@ using NewsAI.Infrastructure.Services;
 
 namespace NewsAI.Data.Service;
 
-public class NewsService : ICommonService<News>
+public class NewsService : INewsService
 {
     private readonly INewsRepository _newsRepository;
-
-    public List<string> Errors { get; } = new List<string>();
 
     public NewsService(INewsRepository newsRepository)
     {
@@ -25,7 +23,9 @@ public class NewsService : ICommonService<News>
 
     public async Task<News?> FindById(Guid id)
     {
-        await ValidateExist(id);
+        var newExist = await ValidateExist(id);
+        
+        if(!newExist) return null;
 
         var news = await _newsRepository.GetNewsByIdAsync(id);
 
@@ -38,25 +38,31 @@ public class NewsService : ICommonService<News>
 
         if (!titleExist) return await _newsRepository.AddNewsAsync(news);
 
-        Errors.Add($"Title already exist");
+        // Errors.Add($"Title already exist");
         return false;
     }
 
     public async Task<bool> Update(News news)
     {
-        var titleExist = await _newsRepository.SearchByTitle(news.Title);
 
-        await ValidateExist(news.Id);
+        var existNew = await ValidateExist(news.Id);
+        
+        if (!existNew) return false;
+        
+        var titleExist = await _newsRepository.SearchByTitle(news.Title);
 
         if (!titleExist) return await _newsRepository.UpdateNewsAsync(news);
 
-        Errors.Add($"Title already exist");
+        // Errors.Add($"Title already exist");
         return false;
     }
 
     public async Task<bool> Delete(Guid id)
     {
-        await ValidateExist(id);
+        
+        var existNew = await ValidateExist(id);
+        
+        if (!existNew) return false;
 
         return await _newsRepository.DeleteNewsAsync(id);
     }
@@ -65,9 +71,6 @@ public class NewsService : ICommonService<News>
     {
         var exist = await _newsRepository.GetNewsByIdAsync(id);
 
-        if (exist != null) return true;
-
-        Errors.Add("No news found");
-        return false;
+        return exist != null;
     }
 }
